nimeche_backend/requirements.txt
asgiref==3.8.1
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-googledrive-storage==1.6.0
djangorestframework==3.15.2
google-api-core==2.10.2
google-api-python-client==2.138.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
